+++ src/JsonApi/JsonApiClient.php
@@ -1,23 +1,23 @@
 <?php
 namespace WoohooLabs\Yang\JsonApi;
 
-use GuzzleHttp\Client;
+use Http\Client\HttpClient;
 use Psr\Http\Message\RequestInterface;
 use WoohooLabs\Yang\JsonApi\Response\JsonApiResponse;
 
 class JsonApiClient
 {
     /**
-     * @var \GuzzleHttp\Client
+     * @var \Http\Client\HttpClient
      */
     protected $client;
 
     /**
-     * @param array $config
+     * @param \Http\Client\HttpClient $client
      */
-    public function __construct(array $config = [])
+    public function __construct(HttpClient $client)
     {
-        $this->client = new Client($config);
+        $this->client = $client;
     }
 
     /**
@@ -26,28 +26,6 @@ class JsonApiClient
      */
     public function request(RequestInterface $request)
     {
-        return new JsonApiResponse($this->client->send($request));
-    }
-
-    /**
-     * @param \Psr\Http\Message\RequestInterface $request
-     * @return \GuzzleHttp\Promise\PromiseInterface
-     */
-    public function requestAsync(RequestInterface $request)
-    {
-        return $this->client->sendAsync($request);
-    }
-
-    /**
-     * @param \Psr\Http\Message\RequestInterface[] $requests
-     * @return \Psr\Http\Message\ResponseInterface[]
-     */
-    public function requestConcurrent(array $requests)
-    {
-        foreach ($requests as $key => $request) {
-            $requests[$key] = $this->client->sendAsync($request);
-        }
-
-        return \GuzzleHttp\Promise\unwrap($requests);
+        return new JsonApiResponse($this->client->sendRequest($request));
     }
 }